Started Frontend

+++ backend/controllers/product.controller.js
@@ -35,7 +35,7 @@ export const updateProduct = async (req, res) => {
     const product = req.body;
 
     if(!mongoose.Types.ObjectId.isValid(id)) {
-        return res.status(400).json({success:false, message: "Invalid product id"});
+        return res.status(404).json({success:false, message: "Invalid product id"});
     }
 
     try {
@@ -50,11 +50,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
     
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({success:false, message: "Invalid product id"});
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success:true, message: "Product deleted"});
     } catch (error) {
         console.error("error in delete product", error.message);
-        res.status(404).json({success:false, message: "Product not found"});
+        res.status(500).json({success:false, message: "Server Error"});
     }
 }
